separate containers and components
