+++ src/components/parser/outputSectionHeader.js
