+++ src/components/parser/parseIcon.js
